ADD - Primera iteracion

+++ src/main/java/ar/edu/undef/fie/criptoboot/entities/Usuario.java
@@ -1,8 +1,12 @@
 package ar.edu.undef.fie.criptoboot.entities;
 
+import lombok.EqualsAndHashCode;
 import lombok.ToString;
 
 import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 
 @Entity
 @ToString
@@ -10,7 +14,7 @@ import javax.persistence.Entity;
 public class Usuario {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
-    private Long id;
+    private int id;
     private String nombre;
     private String apellido;
     private String email;
